# docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task API Documentation",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "https://laravel-sanctum-auth.azurewebsites.net/"
        }
    ],
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Lista todas as tasks",
                "description": "Lista todas as tasks",
                "operationId": "ba6d95ee0baf571660c10b599bcc7855",
                "responses": {
                    "200": {
                        "description": "Retorna uma array de tasks",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Retona a mensagem de erro ou sucesso",
                                            "type": "string",
                                            "example": "Tasks have been retrieved!"
                                        },
                                        "success": {
                                            "description": "Retorna o status os sucesso",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "description": "Retorna o vetor de tasks",
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "description": "ID",
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "description": "título da tarefa",
                                                            "type": "string",
                                                            "example": "tarefa 1"
                                                        },
                                                        "done": {
                                                            "description": "status",
                                                            "type": "boolean",
                                                            "example": false
                                                        },
                                                        "created_at": {
                                                            "description": "status",
                                                            "type": "string",
                                                            "example": "20/12/2023"
                                                        },
                                                        "updated_at": {
                                                            "description": "status",
                                                            "type": "string",
                                                            "example": "20/12/2023"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Acess token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}